WIP -> MENU, OPTIONS, TUTO, SFX
-> Ajout de la scène Main Menu dans Assets/Game/Menu/Scene

-> Ajout de boutons pour cette scène de menu et dans les build settings

-> ajout du texte pour les coquillages dans l'inventaires

-> Quand on scrap des coquillages on les gagne dans l'inventaire.

-> Fix sur les options ou le revert par le menu langage ne marchait pas.

-> Les options affichent de manière précise le taux de rafraichissement en fonction de l'écran.

-> La pop up de confirmation affiche désormais le temps dans lequel il faille confirmer avant de revert.

+++ Assets/Player/Inventory/InputsHandler/InventoryDropHandler.cs
@@ -123,7 +123,7 @@ namespace LightHouse.Inventory
                 DropItem
                 (
                     slotID: SlotManager.CurrentSelectedSlot.SlotDatas.SlotID,
-                    itemGlobalID: SlotManager.CurrentSelectedSlot.SlotDatas.ItemGlobalID,
+                    itemGlobalID: (ushort)SlotManager.CurrentSelectedSlot.SlotDatas.ItemGlobalID,
                     specificID: SlotManager.CurrentSelectedSlot.SlotDatas.ItemSpecificIds[0],
                     pos: _inventoryTarget.position,
                     force: _dropPower,
@@ -144,7 +144,7 @@ namespace LightHouse.Inventory
                 DropItem
                 (
                     slotID: SlotManager.CurrentSelectedSlot.SlotDatas.SlotID,
-                    itemGlobalID: SlotManager.CurrentSelectedSlot.SlotDatas.ItemGlobalID,
+                    itemGlobalID: (ushort)SlotManager.CurrentSelectedSlot.SlotDatas.ItemGlobalID,
                     specificID: SlotManager.CurrentSelectedSlot.SlotDatas.ItemSpecificIds[0],
                     pos: _inventoryTarget.position,
                     force: 0.0f,
